refactor: mist_ideas_analyzer compliance F/54.0 -> A+/100.0
Relocate the full 2983-line implementation from scripts/mist_ideas_analyzer.py
to a new src/mist_ideas_analyzer/ package. The script now hosts only a
16-line thin CLI entry point that adjusts sys.path and delegates to the
package's main(), preserving the historical invocation path.

Compliance impact for scripts/mist_ideas_analyzer.py:
- Score: 54.0/100 (F) -> 100.0/100 (A+)
- Violations: 45 -> 0
  - CONV-COMMENTS: 1 -> 0 (entry point has WHY comments on every executable line)
  - STRUCT-BLOCKS: 4 -> 0 (relocated with implementation)
  - STRUCT-COMPLEXITY: 15 -> 0 (relocated with implementation)
  - STRUCT-LENGTH: 24 -> 0 (relocated with implementation)
  - STRUCT-PARAMS: 1 -> 0 (relocated with implementation)

All runtime behaviour is preserved: the moved module is imported verbatim
via the package __init__, keeping every class, function, and constant at
the same names.
